PM / suspend: Clear wakeups before running PM callbacks
PM callbacks can be used as an indicator that the system is suspending, and
as such, a wakeup may be issued outside the notifier in order to cancel an
ongoing suspend. Clearing wakeups after running the PM callbacks can thus
cause wakeups to be missed. Fix it by simply clearing wakeups prior to
running PM callbacks.

Change-Id: Idf4f4f8d5772d52242151960a092cd81d463095a
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>