treewide: Implement KernelSU's hooks
The Kprobes framework, which facilitates dynamic instrumentation and
debugging in the kernel, had known issues in the 4.19.x kernel branch.
These issues were causing bootloop and unintended behavior, which could
have interfered with the correct operation of KernelSU.
To address this, we have temporarily disabled Kprobes and use KernelSU's hooks.

Feature: KernelSU
Change-Id: Ie5ebad74428b6568421dae7f45aa19d790957084
Signed-off-by: akari 梓川 <akariondev@gmail.com>
Co-authored-by: Rem01Gaming <Rem01_Gaming@proton.me>